move select wallet widgets into separate dir + split into smaller files
