+++ catalyst_voices/apps/voices/lib/pages/registration/wallet_link/stage/select_wallet/widget/wallets_error_state.dart
@@ -1,0 +1,26 @@
+import 'package:catalyst_voices/widgets/widgets.dart';
+import 'package:catalyst_voices_localization/catalyst_voices_localization.dart';
+import 'package:flutter/material.dart';
+
+class WalletsErrorState extends StatelessWidget {
+  final VoidCallback onRetry;
+
+  const WalletsErrorState({
+    super.key,
+    required this.onRetry,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Align(
+      alignment: Alignment.topCenter,
+      child: SizedBox(
+        width: double.infinity,
+        child: VoicesErrorIndicator(
+          message: context.l10n.somethingWentWrong,
+          onRetry: onRetry,
+        ),
+      ),
+    );
+  }
+}
