+++ catalyst_voices/apps/voices/lib/pages/registration/wallet_link/stage/select_wallet/widget/wallets_loading.dart
@@ -1,0 +1,13 @@
+import 'package:catalyst_voices/widgets/widgets.dart';
+import 'package:flutter/material.dart';
+
+class WalletsLoading extends StatelessWidget {
+  const WalletsLoading({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: VoicesCircularProgressIndicator(),
+    );
+  }
+}
